#ticket 2338 comment 111 fix

+++ src/app/dashboard/customerEventView.component.ts
@@ -165,12 +165,15 @@ export class customerEventViewComponent implements OnInit, OnDestroy {
     this.router.navigate(['customerEventList'], { queryParams: { 'force': this.isForce } });
   }
   delete() {
-    this.subscriptions.add(
-      this.administrativeService.deleteEventHistoryById(this.customerId, this.customerEventId)
-      .pipe(skipWhile((item: any) => !item))
-      .subscribe((response: any) => {
-        this.router.navigate(['customerEventList'], { queryParams: { 'force': true } });
-      }));
+    if (confirm('Are you sure you want to delete?')){
+      this.subscriptions.add(
+        this.administrativeService.deleteEventHistoryById(this.customerId, this.customerEventId)
+        .pipe(skipWhile((item: any) => !item))
+        .subscribe((response: any) => {
+          this.router.navigate(['customerEventList'], { queryParams: { 'force': true } });
+        }));
+    }
+    
   }
 
   save() {
